Moved tsx & css files into their own named component file
