+++ src/components/SideBarNav/SideBarNav.tsx
@@ -1,7 +1,7 @@
-import '../styles/SideBarNav.css';
+import './SideBarNav.css'
 import * as React from 'react';
 import { Navbar } from 'react-bootstrap';
-import SideBarButtons from './SideBarButtons';
+import SideBarButtons from '../SideBarButtons/SideBarButtons';
 import { Link } from 'react-router-dom';
 
 interface SideBarNavProps {}
@@ -10,7 +10,7 @@ const SideBarNav: React.FC<SideBarNavProps> = () => {
 	return (
 		<Navbar className="navbar">
 			<Navbar.Brand className="navbar-title">
-				<Link to="/" style={{textDecoration: 'none', color: 'inherit'}}>
+				<Link to="/" style={{ textDecoration: 'none', color: 'inherit' }}>
 					DNVR-<i>zero</i>
 				</Link>
 			</Navbar.Brand>
